group settings in new pages instead of one big page full of settings
