+++ lib/background_service.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_background_service/flutter_background_service.dart';
 import 'package:flutter_background_service_android/flutter_background_service_android.dart';
@@ -48,7 +49,7 @@ onStart(ServiceInstance service) async {
     service.stopSelf();
   });
 
-  Timer.periodic(const Duration(seconds: 10), (timer) {
+  Timer.periodic(kDebugMode ? const Duration(seconds: 10): const Duration(hours: 2), (timer) {
     print('back ground service running');
     _backgroundCheck();
   });
@@ -108,7 +109,7 @@ _handleChannelNotifications() async {
         var locals = await getLocalization();
 
         print('$videosToNotifyAbout videos from channel ${n.channelName} to notify about');
-        if (videosToNotifyAbout >= 0) {
+        if (kDebugMode || videosToNotifyAbout > 0) {
           sendNotification(locals.channelNotificationTitle(n.channelName),
               locals.channelNotificationContent(n.channelName, videosToNotifyAbout),
               type: NotificationTypes.channelNotification, payload: n.channelId, id: n.id);
@@ -155,7 +156,7 @@ _handleSubscriptionsNotifications() async {
         var locals = await getLocalization();
 
         print('$videosToNotifyAbout videos to notify about');
-        if (videosToNotifyAbout > 0) {
+        if (kDebugMode || videosToNotifyAbout > 0) {
           sendNotification(
               locals.subscriptionNotificationTitle, locals.subscriptionNotificationContent(videosToNotifyAbout),
               type: NotificationTypes.subscriptionNotifications);
