+++ lib/myRouteObserver.dart
@@ -5,9 +5,12 @@ import 'package:logging/logging.dart';
 
 const String PATH_CHANNEL = "/channel";
 const String PATH_VIDEO = "/video";
-const String PATH_MANAGE_SUBS = "/manage-subscription_management";
+const String PATH_MANAGE_SUBS = "/manage-subscription-management";
 const String PATH_LAYOUT_EDITOR = '/edit-layout';
 const String PATH_SETTINGS_NOTIFICATIONS = "/settings/notifications";
+const String pathSettingsBrowsing = "/settings/browsing";
+const String pathSettingsVideoPlayer = "/settings/video_player";
+const String pathSettingsAppearance ="/settings/appearance";
 
 const RouteSettings ROUTE_SETTINGS = RouteSettings(name: 'settings');
 const RouteSettings ROUTE_DOWNLOAD_MANAGER = RouteSettings(name: 'download-manager');
