+++ lib/notifications/notifications.dart
@@ -3,9 +3,12 @@ import 'dart:ffi';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:invidious/main.dart';
 import 'package:invidious/myRouteObserver.dart';
+import 'package:logging/logging.dart';
 
 const openChannel = "open-channel";
 const openSubscriptions = "open-subscriptions";
+
+final log = Logger('notifications');
 
 enum NotificationTypes {
   subscriptionNotifications(
@@ -54,11 +57,22 @@ Future<void> initializeNotifications() async {
       onDidReceiveNotificationResponse: onDidReceiveNotificationResponse);
 }
 
+void onNotificationAppLaunch() {
+  // checking if the app was launched from a notification
+  flutterLocalNotificationsPlugin.getNotificationAppLaunchDetails().then((details) {
+    log.info("App started by notification with payload: ${details?.notificationResponse?.payload}");
+    if (details?.notificationResponse != null) {
+      onDidReceiveNotificationResponse(details!.notificationResponse!);
+    }
+  });
+}
+
 void onDidReceiveNotificationResponse(NotificationResponse details) {
   if (details.payload != null && details.payload!.isNotEmpty) {
     var split = details.payload!.split(":");
     switch (split[0]) {
       case openChannel:
+        log.fine('Launching channel screen ${details.payload}, navigator state: ${navigatorKey.currentState}');
         navigatorKey.currentState?.pushNamed(PATH_CHANNEL, arguments: split[1]);
         break;
     }
