+++ lib/settings/views/screens/settings.dart
@@ -30,68 +30,29 @@ class Settings extends StatelessWidget {
   const Settings({super.key});
 
   manageServers(BuildContext context) {
-    Navigator.of(context).push(MaterialPageRoute(
-        settings: ROUTE_SETTINGS_MANAGE_SERVERS,
-        builder: (context) => const ManageServers()));
-  }
-
-  openSponsorBlockSettings(BuildContext context) {
-    Navigator.of(context).push(MaterialPageRoute(
-        settings: ROUTE_SETTINGS_SPONSOR_BLOCK,
-        builder: (context) => const SponsorBlockSettings()));
-  }
-
-  openVideoFilterSettings(BuildContext context) {
-    Navigator.of(context).push(MaterialPageRoute(
-        settings: ROUTE_SETTINGS_VIDEO_FILTERS,
-        builder: (context) => const VideoFilterSettings()));
-  }
-
-  openSearchHistorySettings(BuildContext ctx) {
-    Navigator.of(ctx).push(MaterialPageRoute(
-        settings: ROUTE_SETTINGS_SEARCH_HISTORY,
-        builder: (context) => const SearchHistorySettings()));
+    Navigator.of(context)
+        .push(MaterialPageRoute(settings: ROUTE_SETTINGS_MANAGE_SERVERS, builder: (context) => const ManageServers()));
   }
 
   openAppLogs(BuildContext context) {
-    Navigator.of(context).push(MaterialPageRoute(
-        settings: ROUTE_SETTINGS_SEARCH_HISTORY,
-        builder: (context) => const AppLogs()));
+    Navigator.of(context)
+        .push(MaterialPageRoute(settings: ROUTE_SETTINGS_SEARCH_HISTORY, builder: (context) => const AppLogs()));
   }
 
   openNotificationSettings(BuildContext context) {
     Navigator.of(context).pushNamed(PATH_SETTINGS_NOTIFICATIONS);
   }
 
-  searchCountry(BuildContext context, SettingsState controller) {
-    var locals = AppLocalizations.of(context)!;
-    var cubit = context.read<SettingsCubit>();
-    var colors = Theme.of(context).colorScheme;
-
-    SelectList.show(context,
-        values: countryCodes.map((e) => e.name).toList(),
-        value: controller.country.name,
-        searchFilter: (filter, value) =>
-            value.toLowerCase().contains(filter.toLowerCase()),
-        itemBuilder: (value, selected) => Text(
-              value,
-              style: TextStyle(color: selected ? colors.primary : null),
-            ),
-        onSelect: cubit.selectCountry,
-        title: locals.selectBrowsingCountry);
+  openBrowsingSettings(BuildContext context) {
+    Navigator.of(context).pushNamed(pathSettingsBrowsing);
   }
 
-  String getNavigationLabelText(
-      BuildContext context, NavigationDestinationLabelBehavior behavior) {
-    var locals = AppLocalizations.of(context)!;
-    return switch (behavior) {
-      NavigationDestinationLabelBehavior.alwaysHide =>
-        locals.navigationBarLabelNeverShow,
-      NavigationDestinationLabelBehavior.alwaysShow =>
-        locals.navigationBarLabelAlwaysShowing,
-      NavigationDestinationLabelBehavior.onlyShowSelected =>
-        locals.navigationBarLabelShowOnSelect,
-    };
+  openVideoPlayerSettings(BuildContext context) {
+    Navigator.of(context).pushNamed(pathSettingsVideoPlayer);
+  }
+
+  openAppearanceSettings(BuildContext context) {
+    Navigator.of(context).pushNamed(pathSettingsAppearance);
   }
 
 /*
@@ -112,116 +73,6 @@ class Settings extends StatelessWidget {
     );
   }
 */
-
-  customizeApp(BuildContext context) {
-    showDialog(
-        barrierDismissible: true,
-        context: context,
-        builder: (context) => const AlertDialog(
-            content: SizedBox(width: 300, child: AppCustomizer())));
-  }
-
-  customizeNavigationLabel(BuildContext context) {
-    var locals = AppLocalizations.of(context)!;
-    var settings = context.read<SettingsCubit>();
-    var colors = Theme.of(context).colorScheme;
-    var textTheme = Theme.of(context).textTheme;
-
-    SelectList.show<NavigationDestinationLabelBehavior>(context,
-        values: NavigationDestinationLabelBehavior.values,
-        value: settings.state.navigationBarLabelBehavior,
-        itemBuilder: (value, selected) => Text(
-              getNavigationLabelText(context, value),
-              style: textTheme.bodyLarge
-                  ?.copyWith(color: selected ? colors.primary : null),
-            ),
-        onSelect: settings.setNavigationBarLabelBehavior,
-        title: locals.navigationBarStyle);
-  }
-
-  showSelectLanguage(BuildContext context, SettingsState controller) {
-    var localsList = AppLocalizations.supportedLocales;
-    var localsStrings =
-        localsList.map((e) => e.nativeDisplayLanguageScript ?? '').toList();
-    var locals = AppLocalizations.of(context)!;
-    var cubit = context.read<SettingsCubit>();
-    var colors = Theme.of(context).colorScheme;
-
-    List<String>? localeString = controller.locale?.split('_');
-    Locale? selected = localeString != null
-        ? Locale.fromSubtags(
-            languageCode: localeString[0],
-            scriptCode: localeString.length >= 2 ? localeString[1] : null)
-        : null;
-
-    SelectList.show<String>(context,
-        values: [locals.followSystem, ...localsStrings],
-        value: selected?.nativeDisplayLanguageScript ?? locals.followSystem,
-        itemBuilder: (value, selected) => Text(
-              value,
-              style: TextStyle(color: selected ? colors.primary : null),
-            ),
-        onSelect: (value) {
-          if (value == locals.followSystem) {
-            cubit.setLocale(localsList, localsStrings, null);
-          } else {
-            cubit.setLocale(localsList, localsStrings, value);
-          }
-        },
-        title: locals.appLanguage);
-
-/*
-    SelectDialog.showModal<String>(
-      context,
-      label: locals.appLanguage,
-      selectedValue: selected?.nativeDisplayLanguageScript ?? locals.followSystem,
-      showSearchBox: false,
-      items: [locals.followSystem, ...localsStrings],
-      onChange: (String selected) {
-        if (selected == locals.followSystem) {
-          cubit.setLocale(localsList, localsStrings, null);
-        } else {
-          cubit.setLocale(localsList, localsStrings, selected);
-        }
-      },
-    );
-*/
-  }
-
-  List<String> getCategories(BuildContext context) {
-    var locals = AppLocalizations.of(context)!;
-    return [
-      locals.popular,
-      locals.trending,
-      locals.subscriptions,
-      locals.playlists,
-      locals.history
-    ];
-  }
-
-  selectTheme(BuildContext context, SettingsState _) {
-    var cubit = context.read<SettingsCubit>();
-    var locals = AppLocalizations.of(context)!;
-    showDialog(
-        context: context,
-        useRootNavigator: true,
-        useSafeArea: true,
-        builder: (ctx) => SizedBox(
-              height: 200,
-              child: SimpleDialog(
-                  title: Text(locals.themeBrightness),
-                  children: ThemeMode.values
-                      .map((e) => RadioListTile(
-                          title: Text(cubit.getThemeLabel(locals, e)),
-                          value: e,
-                          groupValue: _.themeMode,
-                          onChanged: (value) {
-                            Navigator.of(ctx).pop();
-                            cubit.setThemeMode(value);
-                          }))
-                      .toList()),
-            ));
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -246,182 +97,75 @@ class Settings extends StatelessWidget {
               lightTheme: theme,
               darkTheme: theme,
               sections: [
-                SettingsSection(
-                  title: Text(locals.browsing),
-                  tiles: [
-                    SettingsTile(
-                      title: Text(locals.country),
-                      value: Text(_.country.name),
-                      onPressed: (ctx) => searchCountry(ctx, _),
-                    ),
-                    SettingsTile(
-                      title: Text(locals.customizeAppLayout),
-                      value: Text(_.appLayout
-                          .map((e) => e.getLabel(locals))
-                          .join(", ")),
-                      onPressed: (ctx) => customizeApp(ctx),
-                    ),
-                    SettingsTile.switchTile(
-                      title: Text(locals.distractionFreeMode),
-                      description: Text(locals.distractionFreeModeDescription),
-                      initialValue: _.distractionFreeMode,
-                      onToggle: cubit.setDistractionFreeMode,
-                    ),
-                    SettingsTile(
-                      title: Text(locals.appLanguage),
-                      value: Text(
-                          cubit.getLocaleDisplayName() ?? locals.followSystem),
-                      onPressed: (ctx) => showSelectLanguage(ctx, _),
-                    ),
-                    SettingsTile.switchTile(
-                      title: const Text('Return YouTube Dislike'),
-                      description: Text(locals.returnYoutubeDislikeDescription),
-                      initialValue: _.useReturnYoutubeDislike,
-                      onToggle: cubit.toggleReturnYoutubeDislike,
-                    ),
-                    SettingsTile.navigation(
-                      title: Text(locals.searchHistory),
-                      description: Text(locals.searchHistoryDescription),
-                      onPressed: (context) => openSearchHistorySettings(ctx),
-                    ),
-                    SettingsTile.navigation(
-                      title: Text(locals.videoFilters),
-                      description:
-                          Text(locals.videoFiltersSettingTileDescriptions),
-                      onPressed: openVideoFilterSettings,
-                    ),
-                  ],
-                ),
-                SettingsSection(title: Text(locals.servers), tiles: [
+                SettingsSection( tiles: [
                   SettingsTile.navigation(
+                    leading: const Icon(Icons.home),
+                    title: Text(locals.browsing),
+                    description: Text(List.of([
+                      locals.country,
+                      locals.customizeAppLayout,
+                      locals.distractionFreeMode,
+                      locals.appLanguage,
+                      "Return YouTube Dislike",
+                      locals.searchHistory,
+                      locals.videoFilters
+                    ]).join(", ")),
+                    onPressed: openBrowsingSettings,
+                  ),
+                  SettingsTile.navigation(
+                    leading: const Icon(Icons.cloud_queue),
                     title: Text(locals.manageServers),
                     description: BlocBuilder<AppCubit, AppState>(
-                        buildWhen: (previous, current) =>
-                            previous.server != current.server,
-                        builder: (context, app) => Text(app.server != null
-                            ? locals.currentServer(app.server!.url)
-                            : "")),
+                        buildWhen: (previous, current) => previous.server != current.server,
+                        builder: (context, app) =>
+                            Text(app.server != null ? locals.currentServer(app.server!.url) : "")),
                     onPressed: manageServers,
-                  )
-                ]),
-                SettingsSection(title: Text(locals.videoPlayer), tiles: [
-                  SettingsTile.switchTile(
-                    initialValue: _.useDash,
-                    onToggle: cubit.toggleDash,
-                    title: Text(locals.useDash),
-                    description: Text(locals.useDashDescription),
-                  ),
-                  SettingsTile.switchTile(
-                    initialValue: _.useProxy,
-                    onToggle: cubit.toggleProxy,
-                    title: Text(locals.useProxy),
-                    description: Text(locals.useProxyDescription),
-                  ),
-                  SettingsTile.switchTile(
-                    initialValue: _.autoplayVideoOnLoad,
-                    onToggle: cubit.toggleAutoplayOnLoad,
-                    title: Text(locals.autoplayVideoOnLoad),
-                    description: Text(locals.autoplayVideoOnLoadDescription),
-                  ),
-                  SettingsTile(
-                    title: Text(locals.subtitleFontSize),
-                    description: Text(locals.subtitleFontSizeDescription),
-                    trailing: Row(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        IconButton(
-                            onPressed: () =>
-                                cubit.changeSubtitleSize(increase: false),
-                            icon: const Icon(Icons.remove)),
-                        Text(_.subtitleSize.floor().toString()),
-                        IconButton(
-                            onPressed: () =>
-                                cubit.changeSubtitleSize(increase: true),
-                            icon: const Icon(Icons.add)),
-                      ],
-                    ),
-                  ),
-                  SettingsTile.switchTile(
-                    initialValue: _.rememberSubtitles,
-                    onToggle: cubit.toggleRememberSubtitles,
-                    title: Text(locals.rememberSubtitleLanguage),
-                    description:
-                        Text(locals.rememberSubtitleLanguageDescription),
-                  ),
-                  SettingsTile.switchTile(
-                    initialValue: _.rememberPlayBackSpeed,
-                    onToggle: cubit.toggleRememberPlaybackSpeed,
-                    title: Text(locals.rememberPlaybackSpeed),
-                    description: Text(locals.rememberPlaybackSpeedDescription),
                   ),
                   SettingsTile.navigation(
-                    title: const Text('SponsorBlock'),
-                    description: Text(locals.sponsorBlockDescription),
-                    onPressed: openSponsorBlockSettings,
+                    leading: const Icon(Icons.video_settings),
+                    title: Text(locals.videoPlayer),
+                    description: Text(List.of([
+                      locals.useDash,
+                      locals.useProxy,
+                      locals.autoplayVideoOnLoad,
+                      locals.subtitleFontSize,
+                      locals.rememberSubtitleLanguage,
+                      locals.rememberPlaybackSpeed,
+                      "SponsorBlock",
+                      locals.lockFullScreenToLandscape,
+                      locals.fillFullscreen
+                    ]).join(", ")),
+                    onPressed: openVideoPlayerSettings,
                   ),
-                  SettingsTile.switchTile(
-                    initialValue: _.forceLandscapeFullScreen,
-                    onToggle: cubit.toggleForceLandscapeFullScreen,
-                    title: Text(locals.lockFullScreenToLandscape),
-                    description:
-                        Text(locals.lockFullScreenToLandscapeDescription),
-                  ),
-                  SettingsTile.switchTile(
-                    initialValue: _.fillFullscreen,
-                    onToggle: cubit.toggleFillFullscreen,
-                    title: Text(locals.fillFullscreen),
-                    description: Text(locals.fillFullscreenDescription),
-                  ),
-                ]),
-                SettingsSection(title: Text(locals.notifications), tiles: [
                   SettingsTile.navigation(
+                    leading: const Icon(Icons.notifications),
                     title: Text(locals.notifications),
                     description: Text(locals.notificationsDescription),
                     onPressed: openNotificationSettings,
                   ),
+                  SettingsTile.navigation(
+                    leading: const Icon(Icons.palette),
+                    title: Text(locals.appearance),
+                    description: Text(
+                      List.of([
+                        locals.useDynamicTheme,
+                        locals.themeBrightness,
+                        locals.blackBackground,
+                        locals.navigationBarStyle
+                      ]).join(", "),
+                    ),
+                    onPressed: openAppearanceSettings,
+                  ),
                 ]),
-                SettingsSection(
-                  title: Text(locals.appearance),
-                  tiles: [
-                    SettingsTile.switchTile(
-                      initialValue: _.useDynamicTheme,
-                      onToggle: cubit.toggleDynamicTheme,
-                      title: Text(locals.useDynamicTheme),
-                      description: Text(locals.useDynamicThemeDescription),
-                    ),
-                    SettingsTile(
-                      title: Text(locals.themeBrightness),
-                      value: Text(cubit.getThemeLabel(locals, _.themeMode)),
-                      onPressed: (ctx) => selectTheme(ctx, _),
-                    ),
-                    SettingsTile.switchTile(
-                      initialValue: _.blackBackground,
-                      onToggle: cubit.toggleBlackBackground,
-                      title: Text(locals.blackBackground),
-                      description: Text(locals.blackBackgroundDescription),
-                    ),
-                    SettingsTile(
-                      title: Text(locals.navigationBarStyle),
-                      value: Text(getNavigationLabelText(
-                          context, _.navigationBarLabelBehavior)),
-                      onPressed: (ctx) => customizeNavigationLabel(ctx),
-                    ),
-                  ],
-                ),
                 SettingsSection(title: (Text(locals.about)), tiles: [
-                  SettingsTile(
-                      title: const Center(
-                          child: SizedBox(
-                              height: 150, width: 150, child: AppIcon()))),
+                  SettingsTile(title: const Center(child: SizedBox(height: 150, width: 150, child: AppIcon()))),
                   SettingsTile(
                     title: Text('${locals.name}: ${_.packageInfo.appName}'),
-                    description:
-                        Text('${locals.package}: ${_.packageInfo.packageName}'),
+                    description: Text('${locals.package}: ${_.packageInfo.packageName}'),
                   ),
                   SettingsTile(
                     title: Text('${locals.version}: ${_.packageInfo.version}'),
-                    description:
-                        Text('${locals.build}: ${_.packageInfo.buildNumber}'),
+                    description: Text('${locals.build}: ${_.packageInfo.buildNumber}'),
                   ),
                   SettingsTile(
                     title: Text(locals.appLogs),
